app/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
email-validator
pytest
